[NUI] Load Tizen.NUI dependencies libs on Windows and MacOS
\### Description of Change ###
This update `.csproj` to automatically define an `MSBuild` compiler
directive `NOTIZEN` when building the project on a MacOS or Windows
systems.
This directive selects which dll names to use when compiling, thus
enabling `Tizen.NUI` to correctly search for libs at Mac and Windows.

This impacts the load of:
- `dali-csharp-binder`
- `capi-appfw-app-manager`
- `capi-appfw-app-common`
- `dlog`

We first tried to define those names at run-time, but
`DllImportAtribbuts` are `C#` `const strings`, which need to be defined
at compile-time, so it raised the need for the preprocessor directive to
select the correct name.

\### API Changes ###
There are no API changes.